# api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ScalexStore Agent Commerce API",
    "version": "1.0.0",
    "description": "Agent-readable commerce endpoints. Read endpoints are static JSON on the CDN; cart/checkout require the commerce server."
  },
  "servers": [
    {
      "url": "https://dailmyshop.netlify.app"
    }
  ],
  "paths": {
...
    "/api/cart/{id}/items": {
      "post": {
        "summary": "Add an item (dynamic)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "product_id",
                  "size"
                ],
                "properties": {
                  "product_id": {
                    "type": "integer"
                  },
                  "size": {
                    "type": "string"
                  },
                  "qty": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}